GraphLaplacian

ManifoldInterpolation.CustomGraph — Type

CustomGraph

This custom datatype allows to save a mesh and the corresponding graph, as well as Metadata of the mesh. The default constructor takes a string to the location of the mesh in a VTK format.

ManifoldInterpolation.computeEigenGraph — Method

computeEigenGraph(graph::CustomGraph,N::Int )

This function computes the smallest N eigenvalues and eigenfunctions.

ManifoldInterpolation.computeEigenGraph — Method

computeEigenGraph(graph::CustomGraph)

This function computes the smallest N eigenvalues and eigenfunctions.

ManifoldInterpolation.createLaplacian — Method

createLaplacian(graph::CustomGraph)

This function creates a graph as well as the corresponding Laplace matrix.

ManifoldInterpolation.distanceEigenfunctions — Function

distanceEigenfunctions(graph1::CustomGraph, graph2::CustomGraph, n::Number=50)

This functions distance between two graphs in 3 norms.

ManifoldInterpolation.loadMesh — Method

loadMesh(graph::CustomGraph)

This function reads the mesh.

ManifoldInterpolation.saveResults — Function

saveResults(graph::CustomGraph, outputPath::String="result")

This function exports the first N smallest eigenvalues as well as the corresponding eigenvectors as a vtu file.

ManifoldInterpolation.transformMesh — Method

transformMesh(graph::CustomGraph, transformationOperator::Function,functionArg)

This functions applies a transformationOperator to the coordinates of the graph.