Interpolation

ManifoldInterpolation.CustomInterpolator — Type

ManifoldInterpolation.SafeInterpolator — Type

SafeInterpolator

Interpolations.interpolate — Function

Interpolations.interpolate — Method

ManifoldInterpolation.exp_log_error — Method

exp_log_error(points, parameter_points, origin, Gr )

This function checks the accuracy of the exp and log maps. This is done by applying first the log map to project a point from the manifold to the tangentspace space. After that this point is retracted back to the manifold. The distance between the new point and the starting point give an indication on how precise the log and exp maps are.

ManifoldInterpolation.initialize — Method

ManifoldInterpolation.interpolate_complete — Method

interpolate_complete( itp_sample_points, itp_sample_points_parameter, itp_parameter_points, origin, n_dim)

This function interpolates between the given sample points. For this it uses the Grassmann manifold and the tangentspace. The sample points correspond to the parameter value where the sample point was sampled. The parameter points are the points where you want new interpolation values. The origin is the point on the Grassmann manifold where the tangentspace is constructed. The number of eigenvectors to interpolate is set by n_dim.

ManifoldInterpolation.interpolate_tangentspace — Method

ManifoldInterpolation.lagrange_basis — Method

ManifoldInterpolation.plot_crossing — Method

plot_crossing(graphs, n_eigenfunctions, n_safety, itp_values, true_values, parameter_points, itp_sample_parameter_point, Gr)

This function allows to quickly plot eigenvalue crossings.

ManifoldInterpolation.plot_interpolation_error — Function

plot_error_interpolation = scatter(title=title, xlabel="Parameter", ylabel="Error",size=(1000,300) ,margin=5Plots.mm)

This function allows for a quick plot of the interpolation error.